Extract scroll logic to viewController
